requirements.txt
accelerate==0.16.0
lpips==0.1.4
Pillow==9.3.0
scipy==1.10.1
tensorboard==2.11.2
torch==1.13.0
tqdm==4.63.1
diffusers[torch]
transformers